[Chore] 박스 컴포넌트 분리
